refactor(api): extract send-order email logic and thin route
