+++ src/shared/lib/emailTemplates.js
